backend/app/schemas/common.py
from datetime import date
from typing import Generic, TypeVar

from pydantic import BaseModel

T = TypeVar("T")


class PaginatedResponse(BaseModel, Generic[T]):
    items: list[T]
    total: int
    page: int
    page_size: int
    pages: int


class ErrorDetail(BaseModel):
    detail: str
    code: str


class DateRange(BaseModel):
    from_date: date | None = None
    to_date: date | None = None
